% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{build.group.labels}
\alias{build.group.labels}
\title{Define a function that writes the group names}
\usage{
build.group.labels(l, font, groupcolor)
}
\arguments{
\item{groupcolor}{}
}
\value{
grouplabels
}
\description{
Define a function that writes the group names
}
