% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{build.group.legend.elements}
\alias{build.group.legend.elements}
\title{Define a function that makes boxes and labels for group type legends}
\usage{
build.group.legend.elements(
  x = 99.208,
  y,
  color,
  width = 6.584,
  height = 6.584,
  label = "group",
  fontsize = 5,
  elementtype = "Branch",
  fontfamily = "'AvenirNext-Bold'"
)
}
\arguments{
\item{fontfamily}{}
}
\value{
legend element
}
\description{
Define a function that makes boxes and labels for group type legends
}
