% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{build.group.legend}
\alias{build.group.legend}
\title{Define a function that builds a legend for group color}
\usage{
build.group.legend(
  yoffset = 0,
  groupslabels,
  groupcolors,
  elementtype = "Branch",
  fontfamily = "'AvenirNext-Bold'"
)
}
\arguments{
\item{fontfamily}{}
}
\value{
group legend
}
\description{
Define a function that builds a legend for group color
}
