% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{build.nodesize.legend}
\alias{build.nodesize.legend}
\title{Define a function that builds a legend for values}
\usage{
build.nodesize.legend(
  yoffset = 0,
  minval,
  maxval,
  minsize,
  maxsize,
  fontfamily = "'AvenirNext-Bold'",
  subtitle = "test"
)
}
\arguments{
\item{subtitle}{}
}
\value{
nodesize legend
}
\description{
Define a function that builds a legend for values
}
