% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{build.text}
\alias{build.text}
\title{Define a function that make a label}
\usage{
build.text(l, labelselect)
}
\arguments{
\item{labelselect}{}
}
\value{
label
}
\description{
Define a function that make a label
}
