% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{build.value.legend}
\alias{build.value.legend}
\title{Define a function that builds a legend for values}
\usage{
build.value.legend(
  yoffset = 0,
  minval,
  maxval,
  palette,
  elementtype = "Branch",
  fontfamily = "'AvenirNext-Bold'",
  subtitle = "test"
)
}
\arguments{
\item{subtitle}{}
}
\value{
legend for values
}
\description{
Define a function that builds a legend for values
}
