% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{color.by.group}
\alias{color.by.group}
\title{Define a function creates color vector from group}
\usage{
color.by.group(df, recolordf, colors, bg.col = "#D3D3D3", categories = NULL)
}
\arguments{
\item{df}{coral tree df}

\item{recolordf}{df to be recolored}

\item{colors}{colors}

\item{bg.col}{background color}

\item{categories}{kinase group}
}
\value{
list of colors
}
\description{
Define a function creates color vector from group
}
