% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{color.by.selected}
\alias{color.by.selected}
\title{Define a function to create a vector of colors based on selected kinases}
\usage{
color.by.selected(df, sel, bg.col, sel.col)
}
\arguments{
\item{df}{coral tree dataframe}

\item{sel}{selected kinases}

\item{bg.col}{background color}

\item{sel.col}{selected color}
}
\value{
vector with color values
}
\description{
Define a function to create a vector of colors based on selected kinases
}
