% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{color.by.value}
\alias{color.by.value}
\title{Define a function creates color vector from values}
\usage{
color.by.value(df, recolordf, colors, heatrange, bg.col = "#D3D3D3")
}
\arguments{
\item{df}{coral tree df}

\item{recolordf}{df to be recolored}

\item{colors}{colors}

\item{heatrange}{range}

\item{bg.col}{background color}
}
\value{
vector with color
}
\description{
Define a function creates color vector from values
}
