% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{convertID}
\alias{convertID}
\title{Define a function the converts input identifiers to coral IDs}
\usage{
convertID(df, recolordf, inputtype)
}
\arguments{
\item{df}{coral tree df}

\item{recolordf}{df to be recolored}

\item{inputtype}{input id type}
}
\value{
recolordf with ids converted to coral.id
}
\description{
Define a function the converts input identifiers to coral IDs
}
