% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{drawrect}
\alias{drawrect}
\title{Define a function that draws a rect}
\usage{
drawrect(x, y, fill, width = 6.584, height = 11.27)
}
\arguments{
\item{height}{}
}
\value{
rectangle
}
\description{
Define a function that draws a rect
}
