% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{make_tree_data}
\alias{make_tree_data}
\title{Compile coral tree}
\usage{
make_tree_data(df, min_col, max_col)
}
\arguments{
\item{max_col}{}
}
\value{
list of coral tree and legend
}
\description{
Compile coral tree
}
