% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{map2color}
\alias{map2color}
\title{Define a function that maps numbers to colors}
\usage{
map2color(
  x = NULL,
  pal = colorRampPalette(c("deepskyblue2", "black", "gold"))(100),
  limits = NULL
)
}
\arguments{
\item{limits}{}
}
\value{
vector of colors
}
\description{
Define a function that maps numbers to colors
}
