% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{resizes.by.value}
\alias{resizes.by.value}
\title{Define a function creates radius vector from values}
\usage{
resizes.by.value(
  df,
  resizedf,
  sizerange,
  controlledrange = FALSE,
  minvalue = 0,
  maxvalue = 5,
  showall = "show"
)
}
\arguments{
\item{df}{coral tree df}

\item{resizedf}{df to be resized}

\item{sizerange}{range of sizes}

\item{controlledrange}{default = FALSE}

\item{minvalue}{default = 0}

\item{maxvalue}{default = 5}

\item{showall}{default = "show"}
}
\value{

}
\description{
Define a function creates radius vector from values
}
