% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORALcli.R
\name{writekinasetree}
\alias{writekinasetree}
\title{Define a function that writes an kinase tree svg file}
\usage{
writekinasetree(svginfo, destination, font, labelselect, groupcolor)
}
\arguments{
\item{groupcolor}{}
}
\value{
tree lines
}
\description{
Define a function that writes an kinase tree svg file
}
